Add ability to "preconfigure" the app
This is done by placing a config.json into the working directory of the
application (with Jetty, this is at the root of the project). The
config.json file should contain an array of configs, each with unique
UUIDs. Note that the UUID field does not enforce constraints, so it
doesn't necessarily actually have to be a UUID...